pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentiment-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.2",
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "torch>=2.5.1",
    "transformers>=4.48.1",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu118" }]
torchvision = [{ index = "pytorch-cu118" }]

[dependency-groups]
dev = ["ruff>=0.11.8"]
